pyproject.toml
[tool.poetry]
name = "market_watch"
version = "0.1.0"
description = "Watching for Market Intelligence"
authors = ["Taylor Curran <taylorcurranagtc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
prefect = ">=2.0b"
beautifulsoup4 = "^4.11.1"
python-dotenv = "^0.20.0"
tweepy = "^4.10.0"
pandas = "^1.4.2"
greenlet = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
